fix: adjust split color function and nested texts
